+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Version 2.0.0
 - Fixed sslCert and sslPrivKey entries being ignored in the config.
 - Changed ssl key creation to be done at startup instead of on npm install.
 - Added error handling during master startup.
+- Factorio game and rcon port now defaults to a random port above 49151.
+- Removed unimplemented mods update command.
+- Fixed rcon password being generated with Math.random().
 
 ### Breaking Changes